5316: move dropdowns out of forms folder to ready the move of forms folder
