+++ assets/admin/components/playlist-drag-and-drop/playlist-drag-and-drop.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { SelectPlaylistColumns } from "../playlist/playlists-columns";
-import PlaylistsDropdown from "../util/forms/multiselect-dropdown/playlists/playlists-dropdown";
+import PlaylistsDropdown from "../util/multiselect-dropdown/playlists/playlists-dropdown";
 import DragAndDropTable from "../util/drag-and-drop-table/drag-and-drop-table";
 import FormCheckbox from "../util/forms/form-checkbox";
 import {
