+++ assets/admin/components/playlist/playlist-form.jsx
@@ -5,7 +5,7 @@ import UserContext from "../../context/user-context";
 import Schedule from "../util/schedule/schedule";
 import { useGetV2TenantsQuery } from "../../../shared/redux/enhanced-api.ts";
 import ContentBody from "../util/content-body/content-body";
-import TenantsDropdown from "../util/forms/multiselect-dropdown/tenants/tenants-dropdown";
+import TenantsDropdown from "../util/multiselect-dropdown/tenants/tenants-dropdown";
 
 /**
  * The playlist form component.
