+++ assets/admin/components/slide/content/station/station-selector.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import MultiSelectComponent from "../../../util/forms/multiselect-dropdown/multi-dropdown";
+import MultiSelectComponent from "../../../util/multiselect-dropdown/multi-dropdown";
 import { displayError } from "../../../util/list/toast-component/display-toast";
 import userContext from "../../../../context/user-context";
