+++ assets/admin/components/util/multi-and-table/select-playlists-table.jsx
@@ -6,7 +6,7 @@ import {
   useGetV2PlaylistsQuery,
   useGetV2PlaylistsByIdSlidesQuery,
 } from "../../../../shared/redux/enhanced-api.ts";
-import PlaylistsDropdown from "../forms/multiselect-dropdown/playlists/playlists-dropdown";
+import PlaylistsDropdown from "../multiselect-dropdown/playlists/playlists-dropdown";
 import { SelectPlaylistColumns } from "../../playlist/playlists-columns";
 import useFetchDataHook from "../fetch-data-hook.js";
